Update README.md
Hey @davidgs ... I LOVE this! So much fun to read and it's such a great project :-) 
I made some very small grammatical edits, but nothing major.
The only thing I would really encourage changing is the elves.... only because we don't yet have a corporate line on questions relating to "automation putting folks out of jobs." I could see that not landing too well with the leadership team! I adjusted the intro to reflect this, but didn't adjust line 227, because it's part of the automated email. Perhaps we could remove the mention of the elves, or say that they're all living their best lives in Thailand too?